util: Added bash scripts to connect and debug app from terminal
wifi connects the device wirelessly. The device should be connected via usb
before running the script. Once the script has been executed successfully,
the device has to be unplugged. When running the command the ip address of the
device has to be passed as argument.

run_app debugs, builds, installs and opens the application in the device.

In order to be able to run the scripts, the directory of the applications has to
be located in the internal file system of the computer where is being developed.
If the directory is in an external hard drive, the scripts won't run because
it's not possible to set permissions in vfat or ntfs file systems.

The scripts have to be executable. To do so run the following commands:
chmod +x wifi
chmod +x run_app

Additionally, run the next command:
chmod +x gradlew

This may be necessary if the gradlew script is not executable.