feat: Added interface to pass data from AsyncTask to Fragment
In order to pass data to ForecastFragment from FetchWeatherTask,
the interface FetchWeatherTaskResponse was created.
That interface is implemented by ForecastFragment and it's an attribute
of FetchWeatherTask. It's passed to the constructor of the AsyncTask.

The function processWeatherData of the interface is called in onPostExecute.
Then, the json data is processed by the fragment.