feat: Added location to settings
Added location preference to settings so the user can get data from a given place.
The default Location is Quito.
The value of the location is passed to the AsyncTask programatically.
SettingsActivity adds the preferences from the xml file.
It also binds the value of the preference to the summary of it.