feat: Added units to settings
Added option to set units to display information. The options are metric and imperial.
The application still retrieves information from openweathermap.org in metric units,
but the application convert those values to the selected system.
Currently, the only values processes are min and max temperature, which are the
only ones parsed from the json string.