fix: Updated ForecastFragment to enable DetailActivity
In commit cd9c82892aae645de74055df7573f4c822ed9784 the function processWeatherData created a new
ArrayAdapter<String> every time. This adapter was set to the ListView and this caused that the
listener for item selection do not work anymore because that listener depends on an adapter
that is gone by then. The solution was to set the adapter as a private variable of the
fragment, then in processWeatherData, simply clear the adapter and add the new items.