feat: Added share option to menu in action bar
Originally DetailActivity and MainActivity had the same menu. However, different
options are needed for the former. To achieve this, DetailFragment was created.
The content of DetailActivity was moved to that fragment. Then, detailfragment.xml
(which is the menu) was inflated in the fragment. Finally, the logic to share the
content of the fragment is under onOptionsItemSelected.

The logic that process the data from MainActivity (when MainActivity starts an intent
to display DetailActivity) was moved to DetailFragment too.