feat: Added logic to retrieve movies data from themoviedb.org
AndroidManifest has permissions for internet and checking network status.
MoviesAsyncTask execute the retrieving process in another thread. It also defines
an interface to be implemented by CatalogFragment which objective is to pass
the data from one thread to another. Since MoviesAsyncTask only takes cares of getting
the raw data, CatalogFragment has the logic to process the json data and update
the adapter. It also checks for network status before executing the async task.
MovieAdapter has updated the url to get the poster of the movie because Movie objects
have only the identifier of the poster. Movie has a new constructor to pass
a string value for the raiting parameter. stringx.xml has the base url to fetch
the data of movies and the poster of them, it also has a message when no connection
is available. styles.xml has updated the style of the movie item to make the poster
visible.