feat: Updated view to GridView
GridView replaced ListView. Styles are set so the posters of the movies
fill the entire screen. Dimensions of the posters are 185 x 277, the height (277)
is used to set the height of the movie item, the scaleType (centerCrop) makes
the image fill all the available space. The image is fully displayed in portrait
mode, however in landscape mode is cut off because of the fixed height (277)
and the scaleTypw (centerCrop).