feat: Added activity to display details of a movie
Movie class was modified to implement the Parcelable interface to pass
data in an intent; a function to get the year of the movie was added.
Listener to handle click on grid item was added to start the activity
with the details of the movie.
The details of the activity are displayed within a fragment, this is
to be able to add a menu options for that specific fragment later.
The fragment has an ScrollView to let the user move the view if there is
no enough room to display all the info; it also has a RelativeLayout
to avoid too many nested views.