feat: Added database structure to persist data
Created a contract and a database helper to persist data in the device.
The objective of the database is to display data when it's not possible
to retrieve data from openweathermap.org either by lack of network connection
or failures in the server side.

There are two tables in the database, one to store information about weather
conditions and another to store information about location. The first database
is linked to the second through the id of the location. The objective of the
location databse is to avoid repetition when storing weather data.