refactor: Modified strings for units
The values of the strings were not modified at all. The array with
the labels and values were moved to arrays.xml. Strings for values
for imperial an metric were defined in strings.xml; the same for
the corresponding labels. Those strings, values and labels, are
referenced in the array under arrays.xml.