refactor: Created ForecastAdapter to bind data to view
ForecastAdapter will be used to bind data using a Loader. The adapter
contains functions formatHighLows and convertCursorRowToUXFormat which
were originally implemented in FetchWeatherTask. It also defines
FORECAST_COLUMNS with the columns to de returned when querying
the database, that parameter is an array of strings that will
be passed as the projection argument in the corresponding functions
Since, that projection will have the fields in a defined position,
indices for each field have been defined, like COL_WEATHER_DATE for
instance. This allows to use those values directly instead of using
Cursor.getColumnIndex.

Utility class has been created to hold util functions used along the
application.