refactor: Modified FetchWeatherTask to make it compatible with Loaders
ArrayAdapter is no longer needed because there is a ForecastAdapter to bind data
to UI, but this new kind of adapter is not required in the task either.
Since the adapter is no longer needed, there is no need to fill it with
data in onPostExecute, so that function is removed. When retrieving data
from the server and inserting it into the database is not necessary to
query that data to verify they were inserted correctly, so the code to
do so under getWeatherDataFromJson has been removed.

Functions formatHighLows and convertContentValuesToUXFormat were removed
since they were implemented in ForecastAdapter in commit 9fea95c. Function
getReadableDateString used in convertContentValuesToUXFormat was also
removed.