refactor: Modified forecast and detail fragments to use Loaders
Fragments were modified to use Loaders to get data to fill the UI.
Each fragment implements LoaderManager.LoaderCallbacks<Cursor>. onStart function
was removed from ForecastFragment to avoid fetching data from openweathermap.org
every time the device is rotated. onLocationChanged fetches data from the service
when the location has changed under settings, this is done calling that function
within onResume.

DetailFragment defines FORECAST_COLUMNS to be passed as a projection when
querying the database. This allows to use defined values for the columns of
a query without using Cursor.getColumnIndex. The ShareActionProvider object
is set with a share intent when the loader has finished loading data (onLoadFinished)
and when creating the menu (onCreateOptionsMenu).