ui: Modified forecast list item
The xml layout was modified to make the list item more appealing.
Style has been applied to order the elements in the item.
It's using linear layout to make easier to sort the items.
Logic bindView in the forecast adapter has been commented out to
avoid errors since the layout is different now.