cleanup: Removed FetchWeatherTask and SunshineService
Those components were removed since their job is now being done by the SunshineSyncAdapter.

Initially, the task were supposed to fetch the data form the server, but the drawback of
that approach is that the task is not tied to the lifecycle of the activity, therefore, there
was a great chance that serveral task were being executed at the same time when the device
is rotated for instance. To solve that, SunshineService were created.

Although, the service solved the problem of the async task, synchronization with the
server was needed, that's why a sync adapter was created, causing the service become
unneeded.