feat: Added URIs for Trailer and Review
The URIs for trailers and reviews have been defined based on their relationships
with movies. There are two kind of movies for each one; one related to the movie,
so the id of the movie is needed, and another without any relationship at all, where
just the id of the traier or review is needed.