refactor: Modified movie details
Interaction with movie to display detailed information of it were removed in commit
d1a60c7 because loaders were implemented to handled data. That feature was re-implemented
but in a different way. Rather than start a new activity and pass a Movie object which
is parcelable, a uri is added to the intent. Then, in the movie details fragment, that
uri is used to query the movie. That fragment also uses a loader to query the
provider.

In addition, constants for projections used when querying the provider were added
in both fragments. The objective of this constants is to get indices references
ready to use to get the values columns after a query has been done.