refactor: Implemented options menu in Activity instead of Fragment
There's no clear reason to implement the options menu in the MainActivity rather than in the CatalogFragment.
In my opinion, the only reason to do that is that the fragment can change and the logic for the menu can
be lost, so to avoid that, it's better to handle the options menu in the activity.