refactor: Reimplemented queryBy in options menu
In commit 5780a6a settings in menu options where implemented. In the activity, the option
to set the queryBy argument was implemented. Then this value was read from SharedPreferences
and passed to the Uri argument in the asyncTask.

In commit a0a81f5 the fragment is attached programatically to the activity. To take advantage
of this, the activity gets the fragment and invoke the function onQueryByChanged, which in turn
updates the catalog of movies by fetching the data from the server using the corresponding
preferred queryBy. This happens in onResume under the activity since MainActivity is paused
when SettingsActivity is called and it resumes after the settings have been changed.