feat: Added reviews of a movie under MovieDetailsActivity
The approach used to accomplish this requires the use of an additional fragment to display
the reviews of the movie. This new fragment along with the fragment to show details are
addee to the acitvity using a FragmentTabHost that allows multiple tabs in the activity,
each one with a corresponding fragment. To switch between tabs, it is necessary to tab
on their corresponding headers.

In the case of the fragment for reviews, it was necessary to use a ViewPager. First,
a custom ViewPager is needed, which enables scrolling since some reviews are quite long
that exceed the viewport. Also, a custom adapter for that pager is needed. This custom
adapter is meant to work with loaders. To read a new review requires to swipe, and since
moving from fragment to fragment is done by tapping, there is no conflict between the
change of fragments in the activity (movie details and movie review fragments) and the
review fragments within MovieReviewFragments.

There is still needed a way to recover the last review when switching to fragments or
rotating the device, in such cases, the app displays the first review (if any) rather
than the previous one.