feat: Added trailers of a movie under MovieDetailsActivity
The approach is similar to the one implemented in 034cd0a, but a bit simpler.
For the trailers, it was not necessary to implement a pager. All the items are
displayed in the same screen, because they're not need that much space to display
the information. When tapping on a intem, a new intent is started to play the
corresponding video. Currently, it only plays videos from youtube.