refactor: Added fragment under Activity for details of movie
The functionality of the application remains the same. The main difference is that the Activity
for detais now contains a fragment, which in turn, contains the element previously located in
the activity. I short, a new layer of abstraction has been added, which objective is to provide
the logic to handle the UI for tablets.

The names of some activities and fragments has been changed to ease the development and the
understanding of the app's structure.