feat: Added support for tablets
In larger devices the application uses two panes to display the list of movies
and the details of one of them in the same screen.

Basically, the application handles this by checking the presence of a container
for the DetailsFragment in activity_catalog.xml, this is stored in a variable which
is used to handle the behaviour accordingly.