feat: Updated logic to avoid deleting previous movies
Previously, every time the app fetched movies from the server, all previous records were removed.
This didn't allow to set movies as favourites because all previous information was deleted.

Now, every time the app fetches movies from the server, it checks to see if the movies are
already in the db and update them accordingly.