feat: Storing movies' posters locally to improve performance
The objective of this feature is to avoid multiple connections to the server and avoid network usage,
and to avoid the problem of the posters not being displayed correctly in the grid. This latter problem
has not been solved because it seems to be something related to grid itself, not to the elements within
it.