ui: Modified styles for details fragment
Recall that details fragments is the main on in this part of the application, which contains
other fragments to handle different set of data. Here, I modified the title to allow just
one line, because in some cases the title is so long that it can not be displayed in the TextView.
I need to add an animation to solve the current situation.