refactor: Fetched trailers using an AsyncTaskLoader
TrailersAsyncTask is implemented as an AsyncTaskLoader. The logic is the same, but now the fragment
does not implement a LoaderCallbacks, but it create two objects of that kind because there are two
loaders needed to retrieve the trailers of the movie. So, basically, the cursor loader is first
initialized, then if no records are found, the async loader is started. After the async loader
finished, the cursor loader is restarted as long as new records have been added to the database.