feat: Added message when there are no reviews available
Unlike movies and trailers, reviews are displayed using a view pager instead of a list view or grid view.
This means that is not possible to set an emptyview. So, the solution is to handle the visibility of a
text view manually, which is done under onLoadFinished in the cursor loader.