refactor: Reduced code repetition when dealing with cursor loaders for reviews and trailers
The functionality is the same, but blocks of code that check for network connectivity,
create a cursor loader, and get movieId out of a uri have been moved to the Utility class.

Also, previously, under onCreateView the movie uri was retrieved just in the case of tablets,
the case for phones was handled under onLoadFinished in the cursor loader. Now, both cases
are handled under onCreateView. The mMovieId attribute is also determined under the same
function, so the loaders do not take care of anything other than fetching and retrieving
content.