fix: Made sure that the fragment is handling reviews of the right movie
When processing the json data for the reviews retrieved from the server, this data comes with and id of
the movie. In order to make sure the reviews are of the right movie a comparison between this id (the
one from the server) and the id held in the fragment is done. Even though it seems not possible that
this situation can happen, it's a good practice to have this.