refactor: Removed unneeded code under loaders for details, reviews, and trailers
In commit 1af671b, the movie uri used to get details, reviews, and trailers of a movie were determined
under onCreateView, freeing the loaders to do that (under onCreateLoader). Now, some code were
doing a discrimination between tablets and phones the get the movie uri, but since the uri was
already determined in onCreateView, part of the code under onCreateLoader for the async and cursor
loader was not necessary and it has been removed.