feat: Added SyncAdapter to connect to the server
Even though it is not really necessary to use a SyncAdapter to connec the server, this implementation
helps to understand how this technology works.

First, there are two key components needed to accomplish the objective: an authenticator, and a sync
adapter. The first one, as the name implies, is used to authenticate the user in the server. In this
case, there is no need to authenticate the user to get data from the server, but, it's still necessary
to implement this element with stub code. The second one is the element that actually connects to the
server to get the data of the movies, it contains the network code to accomplish the mentioned task.

Both elements needs to be bound to the sync adapter framework in order to perform their tasks. To do
so, the corresponding services need to be implemented. Both services are in charge of creating their
corresponding elements, and to bind them to the framework. Since the services are android components,
they need to be declared in the AndroidManifest.xml file. In both cases, intent filters are declared,
which are triggered by the corresponding intent actions. When those filters are triggered, the system
starts the corresponding services.

Finally, since the content provider is going to be attached to the sync adapter, it has to be declared
as syncable in the AndroidManifest.xml file.

Now, this is just the base implementation, the sync adapter still has to be run in order to connect
to the server, which implies that the code that connects to the server needs to be added under the
MoviesSyncAdapter class. This implementation simply allows to add an account in the system related
to the application, which can be seen in settings under accounts sections. The account for popular
movies is already there.