feat: Used SyncAdapter to connec to the server
The first change is to remove the async loader from the fragment.
This causes that the relationship with the cursor loader ends.
So there's no more cursorLoader -> asyncLoader -> cursorLoader
flow. Rather, the new flow is cursorLoader -> syncAdapter, this
means that if there are no records found with the cursorLoader,
then the syncAdapter can sync with the server right away.
Also there's no need to use the variable mFetchFromServerCount
because once the syncAdapter is synced and get the results,
those results are inmediatly added to the database and the
grid view is updated accordingly, so there's no need to restart
the cursor loader.

The second change due to the use of the sync adapter, is that the
code for process the json results gotten from the server are moved
to the sync adapter, freeing the fragment of data processing. The
fragment used to do this because the async task first, and then the
async loader, delivered results which had to be processed by a
receiver which was the fragment.

The sync adapter is only needed for retrieving the movies data, not
the information about reviews and trailers. The reason for this
decision, is that the reviews and trailers depend on a selected movie,
and it's better to get that information on demand using the async loaders.
If sync adapters were used to get this information, then those adapters
would have to make several calls to the server to get the reviews and
trailers of every movie, which is unefficient and drains the battery.