refactor: Replaced async loader with service for trailers
An IntentService has been implemented to get information about
trailers from the service. The benefits of an intent sevice include
stopping the service when work is done, creating a background thread
to perform the tasks, default implementation for onBind and onStartCommand

In this particular case, compare to an async loader, the service does
not provide an noticeable extra benefit. With the async loader, the
flow was cursorLoader -> asyncLoader -> cursorLoader. Using the
service th workflow is similar, but the service replaces the async
loader. Also, a handler needs to be passed from the fragment to
the service, so both components can communicate, and when the
service has finished its work, then the cursor loader can be
re-initialized, which is needed to update the grid view.

The advantage of the service compared to the async loader is that
the code to process the data from the server can be moved from the
fragment to the server itself.