feat: Added meaningful messages when there are no movies
Previously, when there were no movies to display the app showed a message saying that.
Then, the app checked for the network status to see if the lack of connectivity was
the cause of no movies. But, there are other reasons for the lack of movies like the
server is down, the server send invalid data, and others; that information has to be
presented to the users, so they know what is happening.

To accomplish the goal, it is neccesary to define status. These status are defined as
integer flags. They are defined in the sync adapter and set accordingly. One approach
to define these flags is to use enumerated types, but they are inefficient and expensive
compared to the static final integers mentioned above.

The flags are defined using annotations in an interface within the sync adapter. The
annotations used allows to check that the values stored (in the shared preferences)
are one of the defined ones in the interface. Speaking of the shared preferences,
this element is used to store the status and persist that value even if the app
finishes.

The status is set in parts of the sync adapter code where errors can occur, for instance
when the json data received is corrupted or malformed, therefore, not possible to
parse. Then the fragment that started the sync adapter implements the listener that
checks the shared preferences when they have changed. This listener is register in
onResume an unregistered in onPause. Then in the onSharedPreferenceChanged method,
the empty view is updated.

The method to update the empty view checks the status and display the corresponding
message which is defined in strings.

To use the annotations the corresponding library has to be invoked under dependencies
in the gradle file.