feat: Added logic to search movies section
The logic is quite simple. After the search button is tapped, the async loader is
triggered, when the results are fetched and processed which includes adding
records to the db, the cursor loader is started to recover the records based on
the keyword used.

An empty view has been added to show meaningful messages to the users, the
same way as in the catalog section.