feat: Added animation to movie title
The objective of the animation is to display the entire title of the movie
to the user because in some cases the text is so long that is not possible
to display it in the view.

The animation displays the text from left to right, but it would be better
to change the direction from right to left.