feat: Change Actionbar to Toolbar
Modified Catalog and Details activities to use a toolbar rather than an action bar.
This toolbar allows to add new elements to it because it is a view group.

The main task was to create a new style for the toolbar and reference it in the
app's theme. Then, the toolbar was added in the xml layout of the activities, and
programatically set under onCreate. Finally, the settings activity is throwing
an error which needs to be handled.

minSdkVersion has to be upgraded to 15 to 21 to enable features like
android:toolbarStyle and name=android:popupTheme