refactor: Implemented logic for clicking and empty view in recycler view
First, the recycler view has a view holder which is basicallly a quick way to
access to the elements of the view members in the recycler view. Unlike grid
view, recycler view has no method setOnItemClickListener, for that reason
tha view holder implements the interface View.OnClickListener. Remember that
the view holder gives access to the elements of each member in the recycler.

Then, the adapter itself has a member interface to handle the clicks. This
interface has a method onClick that is called within the onClick method of
the view holder, which is implementing View.OnClickListener.

The constructor of the adapter now needs to pass the interface to handle the
clicks. Since an iterface is passed, it has to be implemented when the
constructor is called in the fragment (onClick method). So, in the fragment
that uses the adapter, the code within GridView.setOnItemClickListener is
moved to the implementation of the interface in the constructor.

In summary, the fragment stills implements the logic for clicking a movie,
this logic is passed to the adapter, which in turns passes it to the view
holder which executes it.

Unlike the grid view, the recycler neither receives an empty view. To get
this feature working with this element, it has to be passed to the adapter
in the constructor. Then, the adapter handles the visibility of this element
by checking the count of elements in the cursor when swapping cursors.
Updating the messages of the empty view is still handled by the fragment.