fix: Solved problem is incorrect update in the recycler view
The problem was that the grid view then, and now the recycler view, was not able
to update its elements properly when scrolling. It seems that the problem has
nothing to do with the view or its adapter, and the source of the issue is either
the loader or the content provider.

In any case, the workaround was to use RecyclerView.setItemViewCacheSize(int).
This method completely removes the problem, but it may arise a performance issue.