feat: Added title and default image for movies
Added the title of the movie below its poster. This redundancy is helpful since
some movies do not have any poster. Also, a placeholder is used until the poster
of the movie is loaded, and to fill the space when there is no poster.

Placeholder image by Freepik http://www.flaticon.com/free-icon/film-reel_129461