feat: Added transitions to and from details activity
The transitions are easier to use since v21, that's why they're are put in the corresponding
folder. When dealing with transitions between activities, those activities have to be able
to respond properly, which is done in the styles. In this case, there are an individual theme
for each activity, inheriting from the main style. The styles of each activity are assigned
in the manifest file.

Then, when starting the activity in the single pane mode the corresponding code is modified
to enable the transitions.