feat: Started fragment in catalog activity base on previous usage
When the Spinner was added to replace the FragmentTabHost to switch between search
and catalog fragment, the later was added twice, first by checking the saved instance
state, and then by the spinner because the item selected action is fired as soon as
the spinner is created. To solve the problem, the creation of the fragment is done
just by the spinner.

Along with the previous fix, the creation of the fragment is based on previous usage
because when switching fragments by using the spinner, a value is stores in shared
preferences, so this will help to recover the previous state and display the
corresponding fragment, which is more user friendly.