feat: Added shared element when transitioning to movie details
In order to make the application more appealing, the movie poster is shared between the catalog
and details activity. Since the activities have fragments, it is necessary to hold or pause
the transitions until the view elements are loaded and laid out. Basically, the transitions are
pause when the activity is created, and they resume when the data is loaded, under onLoadFinished.