fix: Fixed bug found by Jamal
The application checks if the data of a stock is valid. Unfortunately, yahoo
api replies with a valid json string even if the stock is not valid. However,
a stock can be checked by verifying its fields. When a stock is not valid,
all its fields are set to null except symbol and Symbol. This is what the
application does, it checks the fields of the stock. If the stock is invalid,
then there is no need to insert into the content provider.