feat: Added meaningful messages to user
In order to let the user know what is going on when something unexpected has
happened in the application, meaningful comments have been added to the app.

First, app status variables have been defined, these values are used to set an
shared preference that holds this value for further use. The majority of errors
are handled by try/catch blocks, so when an exception is caught, then the status
of the app is set accordingly. Most of these errors occur in the services.

In order to show the messages to the user a text view has been added to the
layout. Once an operation has finished, that view is updated accordingly.

The stock that is going to be queried is stored as a shared preference, this
value is going to be used in the toast presented to the user when that stock is
not valid.