refactor: Removed service handler
The objective of that handler was to communicate with the ui once the service
finishes its work, so that the ui knows that and restart the loader to update
the list of stocks. However, that is not needed because the adapter of the
recycler view has a data observer that notifies changes in the data set.
This way the recycler view is updated properly.