feat: Modified items in recycler view to display stock right away when added
When adding new stocks, the application started the service to do the work, however,
that process takes some time and the user may not know what is happening. So, when
adding the stock, it is inserted into the content provider right away, so it is
shown to the user. While the service is querying the data for the stock, the item
displays a progress bar which is removed and replaced with data when the service
has finished its job. If the stock is not valid, then the record is deleted from
the content provider, and the corresponding item is removed from the view.