feat: Updated the stocks records when the activity is created
The objective of this feature is to let him know what is happening with
the application. As soon as the app starts, the records are updated so
they are temporary and not current. This makes the app displays the
progress bar in every item in the recycler view. Then, the server is
queried to get new data. After the operation is done, the records are
updated, either with new data (if new data was retrieved from the server)
or with previous data (when no data was retrieved from server). When
new data is fetched, the text in the items of the recycler view are bold,
otherwise they are italic.