refactor: Modified error messages to use a toast
Since the application will display information as long as there are data in
the internal storage, that means that the text view for message will be
visible only when no stock records are stored in the app. So, for any other
kind of error, it is better to show a display with the corresponding message.