feat: Added visual alerts to line graph activity
An text view for message and a progress bar have been added. The trick is the following:
The loader is restarted every time a change in the dataset
happens. When the application first starts, it gets the data from the
content provider. However, that data may be not the most recent, or
there may not be data at all. So the server is queried, when this task
finishes, new data may have been collected or not. In either case the
activity needs to get notified to update the ui properly. The special
case happens when there is no data in the content provider, and no
data were retrieved from the server. In this case, the data from
the content provider has not been modified, hence the loader is
not restarted. To allow the loader to be restared, no matter the
case, a dumb record is added at the beginning of the server
transaction, and it is deleted at the end. This solves the problem
of the ui being updated properly. However, the activity needs to
know if the transaction to the server is still in progress, that's
why the shared preference querying_historic is used.