fix: Checked network availability on demand
In MyStocksActivity, network availability was checked in onCreate.
That value was stored in a variable. This worked when the activity
was created, but in the case that the network was not available when
the activity was created, and later the network was available, the
stored value was still false. So instead of storing that value at
start, I better determine it when it is needed.