fix: Fixed update the line graph
The line graph was not being updated properly. I made sure the cursor is
in the first record before iterating it in the adapter.

Also, in the stock intent server, the loop that checks for historic data
was part of the result loop, it has to be independent for the ui to
be updated properly.