feat: Added real data to the widget
To do so, I created an intent service which will be started under onUpdate
in the widget provider. This allows to do the operations to update the data
in a background thread.

The task service is set to do periodic updates of the data in the stocks
activity. Taking advantage of this, the widgets are updated when the service
finishes its work. When this happens a broadcast intent is sent. Since the
widget provider extends BroadcastReceiver, it can listens this kind of intents
and acts accordingly (which means update the widget). To be able to listen to
this kind of broadcast intents, the widget provider has to be set in the
manifest file with the corresponding filter.