feat: Added collection widget to display all the stocks
The collection widget allow to displays all the stocks at once using a list.
Clicking on the header opens the application in the my stocks activity, whereas
clicking in a stocks opens it in the graph line activity.

The clicking in the header behaviour is simply done using a pending intent.
To handle the behaviour for clicks in the stock, a pendint intent template
and a fill intent are needed. The first one sets the commo parts shared by
every element, that is the activity to be opened. This is set under the widget
provider. The fill in intent handles the unique parameters for each stock
item, which in the current case is the uri for that stock, which will be used
to restore the corresponding data for the stock.

Similarly to the simple widget, this collection widget uses a service to fill
its views with data. However, the simple widget just starts an intent service,
whereas the collection widget needs an adapter to bind the data and create a
list item for each row. This adapter is a RemoteViewsService, when it is set to
the list view, the lifecycle of the remote views factory (provided by the
service) is triggered and the data is reloaded and the corresponding item is
created for each record.

The widget is updated by receiving the data broadcasted, but instead of calling
a new intent service, notifyAppWidgetViewDataChanged is called, which triggers
onDataSetChanged, which queries for new data.