fix: Added workaround to update line graph
When updating the line graph with new historic data, there was an error
thrown because of the number of points added to the chart. To avoid this,
I created an array of Charset with the LineSet so the corresponding add
method is called (the one that does not throw any error).