feat: Showed existing stock even though some information is missing
Previously, when an existing stock didn't have all the data to be displayed, the
application did not handle that stock, meaning that it used to be thrown away.
However, to make the application more user friendly, it is better to display the
available data and let the user know that some data is missing.

To accomplish that objective, the application stores the stock data even though
some fields may be missing, and notifies the user about the situation. This is done
by throwing an IncompleteDataStockException; when it is caught, the application
saves its status with the corresponding value.

This new behaviour leads to a situation where even though a task service has been
able to complete, some errors may have occured, but they should be handled
properly. In summary, the application must not have to set its status to ok
after a service is done. Rather, the status should be reset to OK, before
doing any kind of task, and the value should be updated accordingly whenever
an issue is handled.

Finally, the InvalidStockException has been renamed to NonExistingStockException
to be more precise about what is going on.