refactor: Updated error handling for json parsing and non existing stocks
There are two kind of possible errors when handling data from server. The first
one is related to reading the stream from the server, which is mapped to INVALID_DATA.
The second one is related to json parsing, that is after the stream has been
successfully read, the result may not be a well formed json set of data, this is
mapped to INVALID_JSON_DATA.

In commit 3e3f68d, InvalidStockException was renamed to NonExistingStockException,
that error is now mapped to the corresponding status NON_EXISTING_STOCK.

The strings related to the status have been updated accordingly.

Finally, meaningful messages to the developer have been added when catching the
exceptions handled by the application.