feat: Added free and paid versions of the app
The difference between the two versions is the ads, they do not exist in the paid
version. Since the ads are extra content in the application, the paid version is the
basic version. This means that the free version has extra features (the ads).

Based on the previous, the paid version uses the code and resources of the main
version. There is no much to do in this version, other than setting the label of
the application in strings.xml.

In the other hand, the free version adds more features (the ads). That means
that more logic and resources have to be created for this version. In this case,
the layouts for the activity and layout have to be redefined here because of the
extra content, and the context. The MainActivityFragment is redefined here to
enable the ads.

The free version redefines the android manifest to be able to use the resources
for ads. Also, the dependency for ads is only added in the free version.